requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (SQLite — built-in, no driver needed)
sqlalchemy>=2.0.36
alembic>=1.13.0

# Vector Store
faiss-cpu

# Embeddings
sentence-transformers

# LLM (Groq)
groq==0.4.2
httpx==0.27.2

# Document Processing
PyMuPDF
python-docx==1.1.0

# Utilities
pydantic-settings==2.1.0
python-dotenv==1.0.0
tiktoken
tenacity==8.2.3

# Frontend serving
jinja2==3.1.2
aiofiles==23.2.1
